Changed to processing effects instead of frames
